courses/CS101/assignments/hw1/template/requirements.txt
numpy>=1.26,<2.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0


